LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__Http.https.tentative-expected.txt


FAIL __Http: Does not set via DOM 'Path=/' assert_equals: `__Http-prefixtestcookie=0.580765634561275` in `document.cookie` expected false but got true
FAIL __Http: Does not set via DOM 'Secure; Path=/' assert_equals: `__Http-prefixtestcookie=0.8232361667392922` in `document.cookie` expected false but got true
PASS __Http: Does not set via DOM 'Secure; Path=/; httponly'
FAIL __Http: Does not set via HTTP with 'Path=/;' (without Secure) assert_equals: expected (undefined) undefined but got (string) "0.8347201299426504"
FAIL __Http: Does not set via HTTP with 'Secure; Path=/' assert_equals: expected (undefined) undefined but got (string) "0.7890718398120622"
PASS __Http: Does set via HTTP with 'Secure; Path=/;httponly'
PASS __Http: Does set via HTTP with 'Secure; Path=/cookies/;httponly'
